requirements.txt
pytorch
numpy
yaml
pil
scipy
sklearn
matplpotlib